backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cassandra-driver==3.29.1
cassio==0.1.5
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.1
dataclasses-json==0.6.4
distro==1.9.0
easyocr==1.7.1
etelemetry==0.3.1
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastapi==0.110.1
filelock==3.13.3
fitz==0.0.1.dev2
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.3.1
geomet==0.2.1.post1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
imageio==2.34.0
InstructorEmbedding==1.0.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.40
lazy_loader==0.3
looseversion==1.3.0
lxml==5.2.1
MarkupSafe==2.1.5
marshmallow==3.21.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.2.1
ninja==1.11.1.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
openai==1.16.2
opencv-python-headless==4.9.0.80
orjson==3.10.0
packaging==23.2
pandas==2.2.1
pathlib==1.0.1
pillow==10.3.0
prov==2.0.0
pyclipper==1.3.0.post5
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
shapely==2.0.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
traits==6.3.2
transformers==4.39.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.1
yarl==1.9.4
pymupdf

